WIP: Improve buffer alignment in OctreeNode .
This addresses warnings like:

/src/ITK/Modules/Core/Common/src/itkOctreeNode.cxx:67:74:
warning: cast from 'char*' to 'itk::OctreeNodeBranch*'
increases required alignment of target type [-Wcast-align]

with ARM.

The ColorTable is now an array of OctreeNodeBranch instead
of an array of char.

Note: The data in this array isn't actually used, it's a
placeholder to indicate whether a node is a color or a
sub-node (OctreeNodeBranch).

Change-Id: If53cf1d771eab0f81a61c49df7443ea362988886